add run control to maintenance_add.sh

+++ src/main/script/maintenance_add.sh
@@ -7,6 +7,9 @@ dataset="wikidata"
 
 MAX_HOPNUM=$(get_MAX_HOPNUM $dataset)
 
+run_safe=1
+run_naive=0
+
 cur_dir="/hdd/code/yuhansun/data/${dataset}/add"
 backup_dir="${cur_dir}/backup"
 
@@ -39,42 +42,45 @@ add_edge_path="${cur_dir}/edges.txt"
 test_count=1000
 
 ##### No safe nodes strategy #####
-safe_nodes_path="${cur_dir}/safeNodes.txt"
-output_path="${result_dir}/${dataset}"
+if [ $run_naive == 1 ]; then
+	safe_nodes_path="${cur_dir}/safeNodes.txt"
+	output_path="${result_dir}/${dataset}"
 
-# remove current db_dir and use the backup to restore
-rm -r ${db_dir}
-cp -a ${backup_dir}/neo4j-community-3.4.12_${suffix}_after_pn $db_dir
+	# remove current db_dir and use the backup to restore
+	rm -r ${db_dir}
+	cp -a ${backup_dir}/neo4j-community-3.4.12_${suffix}_after_pn $db_dir
 
-rm $safe_nodes_path
-cp ${backup_dir}/safeNodes.txt $safe_nodes_path
+	rm $safe_nodes_path
+	cp ${backup_dir}/safeNodes.txt $safe_nodes_path
 
-java -Xmx100g -jar ${jar_path} \
-	-f addEdgeExperiment \
-	-dp ${db_path}	\
-	-MAX_HOPNUM ${MAX_HOPNUM}	\
-	-maxPNSize ${maxPNSize}	\
-	-edgePath ${add_edge_path}	\
-	-queryCount ${test_count}	\
-	-safeNodesPath ""	\
-	-outputPath ${output_path}
-
+	java -Xmx100g -jar ${jar_path} \
+		-f addEdgeExperiment \
+		-dp ${db_path}	\
+		-MAX_HOPNUM ${MAX_HOPNUM}	\
+		-maxPNSize ${maxPNSize}	\
+		-edgePath ${add_edge_path}	\
+		-queryCount ${test_count}	\
+		-safeNodesPath ""	\
+		-outputPath ${output_path}
+fi
 
 ###### Safe nodes strategy used #######
-output_path="${result_dir}/${dataset}_with_safe"
+if [ $run_safe == 1 ]; then
+	output_path="${result_dir}/${dataset}_with_safe"
 
-# remove current db_dir and use the backup to restore
-rm -r ${db_dir}
-cp -a ${backup_dir}/neo4j-community-3.4.12_${suffix}_after_pn $db_dir
+	# remove current db_dir and use the backup to restore
+	rm -r ${db_dir}
+	cp -a ${backup_dir}/neo4j-community-3.4.12_${suffix}_after_pn $db_dir
 
-java -Xmx100g -jar ${jar_path} \
-	-f addEdgeExperiment \
-	-dp ${db_path}	\
-	-MAX_HOPNUM ${MAX_HOPNUM}	\
-	-maxPNSize ${maxPNSize}	\
-	-edgePath ${add_edge_path}	\
-	-queryCount ${test_count}	\
-	-safeNodesPath $safe_nodes_path	\
-	-outputPath ${output_path}
+	java -Xmx100g -jar ${jar_path} \
+		-f addEdgeExperiment \
+		-dp ${db_path}	\
+		-MAX_HOPNUM ${MAX_HOPNUM}	\
+		-maxPNSize ${maxPNSize}	\
+		-edgePath ${add_edge_path}	\
+		-queryCount ${test_count}	\
+		-safeNodesPath $safe_nodes_path	\
+		-outputPath ${output_path}
+fi
